+++ tests/modules/source_repository/test_plan.py
@@ -12,23 +12,49 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import pytest
 
-
-@pytest.fixture
-def resources(plan_runner):
+def test_resource_count(plan_runner):
+  'Test number of resources created.'
   _, resources = plan_runner()
-  return resources
+  assert len(resources) == 1
 
 
-def test_resource_count(resources):
-  "Test number of resources created."
-  assert len(resources) == 2
+def test_iam(plan_runner):
+  'Test IAM binding resources.'
+  group_iam = '{"fooers@example.org"=["roles/owner"]}'
+  iam = '''{
+    "roles/editor" = ["user:a@example.org", "user:b@example.org"]
+    "roles/owner" = ["user:c@example.org"]
+  }'''
+  _, resources = plan_runner(group_iam=group_iam, iam=iam)
+  bindings = {
+      r['values']['role']: r['values']['members']
+      for r in resources
+      if r['type'] == 'google_sourcerepo_repository_iam_binding'
+  }
+  assert bindings == {
+      'roles/editor': ['user:a@example.org', 'user:b@example.org'],
+      'roles/owner': ['group:fooers@example.org', 'user:c@example.org']
+  }
 
 
-def test_iam(resources):
-  "Test IAM binding resources."
-  bindings = [r['values'] for r in resources if r['type']
-              == 'google_sourcerepo_repository_iam_binding']
-  assert len(bindings) == 1
-  assert bindings[0]['role'] == 'roles/source.reader'
+def test_triggers(plan_runner):
+  'Test trigger resources.'
+  triggers = '''{
+    foo = {
+      filename = "ci/foo.yaml"
+      included_files = ["**/*yaml"]
+      service_account = null
+      substitutions = null
+      template = {
+        branch_name = null
+        project_id = null
+        tag_name = "foo"
+      }
+    }
+  }'''
+  _, resources = plan_runner(triggers=triggers)
+  triggers = [
+      r['index'] for r in resources if r['type'] == 'google_cloudbuild_trigger'
+  ]
+  assert triggers == ['foo']