+++ CHANGELOG.md
@@ -13,6 +13,8 @@ All notable changes to this project will be documented in this file.
 - add support for Cloud Source Repositories in stage 0 and 1 CI/CD
 - fix Gitlab workflow indentation
 - remove unsupported attributes and add supported ones to the Gitlab mapping used for Workload Identity Federation pools
+- add roles for CI/CD source repositories to stage 1 service account on automation project
+- fixes to CI/CD source repositories in stage 1
 
 ## [16.0.0] - 2022-06-06